Remove mcdMigration upgrade scripts
