+++ upgrade/.env.example
@@ -15,8 +15,6 @@ ROLE=confirm
 
 
 # Optional parameters, depends on the script used
-NEW_IMPLEMENTATION_XDAI_BRIDGE=
-NEW_IMPLEMENTATION_ETH_VALIDATORS=
 NEW_IMPLEMENTATION_ETH_BRIDGE=
 HOME_DAILY_LIMIT=100000000000000000000001
 CHAI_INTEREST_RECEIVER=
